Refactor app into modules and split tests
